Added Policies

+++ app/Http/Controllers/BrandController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreBrandRequest;
 use App\Http\Resources\BrandResource;
 use App\Models\Brand;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
 class BrandController extends Controller
@@ -16,6 +17,7 @@ class BrandController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Brand::class);
         return  BrandResource::collection(Brand::all());
     }
 
@@ -27,6 +29,7 @@ class BrandController extends Controller
      */
     public function store(StoreBrandRequest $request)
     {
+        $this->authorize('create', Brand::class);
         $newBrand = Brand::create($request->validated());
         return new BrandResource($newBrand);
     }
@@ -39,6 +42,7 @@ class BrandController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view', Brand::class);
         return new BrandResource(Brand::find($id));
     }
 
@@ -49,8 +53,9 @@ class BrandController extends Controller
      * @param int $id
      * @return BrandResource
      */
-    public function update(StoreBrandRequest $request, $id)
+    public function update(Request $request, int $id)
     {
+        $this->authorize('update', Brand::class);
         $existingBrand = Brand::find($id);
         $existingBrand->update($request->all());
         return new BrandResource($existingBrand);
@@ -64,9 +69,10 @@ class BrandController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Brand::class);
         Brand::destroy($id);
-        if (! Brand::find($id)) {
-            return "Brand deleted successfuly";
+        if (!Brand::find($id)) {
+            return "Brand deleted successfully";
         } else {
             return "Brand not found";
         }
